sets up routing
