+++ src/App.jsx
@@ -1,16 +1,18 @@
 import React from 'react';
+import {Routes, Route} from 'react-router-dom';
 import './App.css';
-import Todos from './components/Todos';
 import Navbar from './components/Navbar';
+import Todos from './pages/todos';
+import { TodoDetails } from './pages/todos/details';
 
 function App() {
   return (
     <div className="App">
       <Navbar/>
-      <div className='container p-4'>
-         <Todos></Todos>
-      </div>
-    
+      <Routes>
+        <Route path='/' element={<Todos></Todos>}/>
+        <Route path='/todos/:todoId' element={<TodoDetails></TodoDetails>}/>
+      </Routes>
     </div>
   );
