+++ src/pages/todos/details/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const TodoDetails = () => {
+  return (
+    <div>TodoDetails</div>
+  )
+}
