+++ src/pages/todos/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import TodosList from '../../components/TodosList'
+
+const Todos = () => {
+  return (
+    <div><TodosList/></div>
+  )
+}
+
+export default Todos
